Extract `storage::from_environment()` fn
